+++ src/app/components/parts/allele-freq/allele-freq.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { TableService } from '../../../services/column-service';
+import { TableService } from '../../../services/table-service';
 
 @Component({
     selector: 'app-allele-freq',
