+++ src/app/components/parts/variants-table/variants-table.component.spec.ts
@@ -22,7 +22,7 @@ import { MockVsalService } from '../../../mocks/vsal-service.mock';
 import { EnsemblService } from '../../../services/ensembl-service';
 import { OverlayMenuComponent } from '../overlay-menu/overlay-menu.component';
 import { ColumnsMenuComponent } from '../columns-menu/columns-menu.component';
-import { TableService } from '../../../services/column-service';
+import { TableService } from '../../../services/table-service';
 import { FilterAutoComponent } from '../filter-auto/filter-auto.component';
 import { RegionInformationComponent } from '../region-information/region-information.component';
 import { ClincalFilteringComponent, ClinicalChart } from '../clincal-filtering/clincal-filtering.component';
@@ -62,7 +62,9 @@ describe('VariantsTableComponent', () => {
                 {
                     provide: TableService,
                     useValue: {
-                        activeColumns: (): any[] => []
+                        activeColumns: (): any[] => [],
+                        sortMap: {},
+                        keys: () => []
                     }
                 },
                 {
@@ -82,15 +84,7 @@ describe('VariantsTableComponent', () => {
                 {
                     provide: VsalService,
                     useValue: new MockVsalService([])
-                },
-                {
-                    provide: EnsemblService,
-                    useValue: {
-                        healthCheck: () => {
-                            return Promise.resolve();
-                        }
-                    }
-                },
+                }
             ]
         })
             .compileComponents();
