+++ src/app/services/filter.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { FilterService } from './filter.service';
-import { TableService } from './column-service';
+import { TableService } from './table-service';
 
 describe('FilterService', () => {
   beforeEach(() => {
